==> Step_2_Label_Encoding_Process[1].ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
import numpy as np

## Label Encoding

df_new =pd.read_csv('/content/drive/MyDrive/CSCT MASTER PROJECT/clean_crop.csv')

df_new

df_new = df_new.drop(['Unnamed: 0'], axis=1)

df_new

## Selecting non-numeric data columns for further processing
agri_obj = df_new.select_dtypes(object) #non numeric data

agri_obj

# Selecting numeric data columns for further analysis
agri_num = df_new.select_dtypes(["int64","float64"]) # numeric data

agri_num

agri_num

# Creating a copy of the original DataFrame for data encoding
df_encoded = df_new.copy()

# Initialize the LabelEncoder
le = LabelEncoder()

# Encoding categorical variables using Label Encoding
# Iterating through each column containing non-numeric data in 'agri_obj'
# and transforming it into numerical format in the 'df_encoded' DataFrame
for col in agri_obj.columns:
    df_encoded[col] = le.fit_transform(agri_obj[col])

# Encoding categorical variables using Label Encoding
# Applying Label Encoding to each column containing non-numeric data in 'agri_obj'
# and updating the corresponding columns in 'df_encoded'
df_encoded[agri_obj.columns] = agri_obj.apply(lambda col: le.fit_transform(col))

# Display the first few rows of the encoded DataFrame 'df_encoded' to inspect the encoded categorical data
print(df_encoded.head())

# Save the encoded DataFrame 'df_encoded' to a CSV file named 'df_encoded.csv' without including the index
df_encoded.to_csv('/content/drive/MyDrive/CSCT MASTER PROJECT/df_encoded.csv', index=False)

# Read the encoded DataFrame from the 'df_encoded.csv' CSV file and store it in a new DataFrame 'df_new'
df_encoded_final= pd.read_csv('/content/drive/MyDrive/CSCT MASTER PROJECT/df_encoded.csv')

df_encoded_final
