==> Step_3_Train-Test_Data_Data_modelling[1].ipynb <==
import pandas as pd
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_squared_error, r2_score, mean_absolute_error
from sklearn.tree import DecisionTreeRegressor
from sklearn.metrics import mean_squared_error,mean_absolute_error, r2_score
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from sklearn.pipeline import Pipeline
import numpy as np
import joblib

# Read the encoded DataFrame from the specified file path and store it in a DataFrame named 'train_test_df'
train_test_df = pd.read_csv('/content/drive/MyDrive/CSCT MASTER PROJECT/df_encoded.csv')

# Get the data types of each column in the 'train_test_df' DataFrame and store them in 'training_data_dtypes'
training_data_dtypes = train_test_df.dtypes

# Display the data types of columns in the training data
print("Training Data Column Data Types:")
print(training_data_dtypes)

# Create a DataFrame 'features' by dropping specified columns
# - 'Unnamed: 0': An unnamed column that may represent an index; not needed as a feature.
# - 'Production' and 'Production Units': These columns contain total production data, which is not used as a feature for yield prediction.
features = train_test_df.drop(['Production','Production Units'], axis=1)

# Define target variable 'Yield'
target = train_test_df['Yield']

# Split the dataset into training and testing subsets.
# - 'features': The input features for the machine learning model.
# - 'target': The target variable (crop yield) that the model aims to predict.
# - 'test_size': The proportion (25%) of the dataset to use for testing, with the remainder used for training.
# - 'random_state': A seed value (42) for the random number generator, ensuring reproducibility of the split.
X_train, X_test, y_train, y_test = train_test_split(features, target, test_size=0.25, random_state=42)

##Linear Regression Model

# Create a machine learning pipeline for the Linear Regression model.
# - The pipeline is defined as a sequence of processing steps.
# - In this case, the only step included is 'model', which represents the Linear Regression model.
# - Additional steps can be added here for data preprocessing, feature engineering, etc., as needed.
pipeline = Pipeline([
    ('model', LinearRegression())  #
])

print(X_train.dtypes)

# Fit (train) the machine learning pipeline on the training data.
# - This step trains the Linear Regression model on the training dataset (X_train and y_train).
# - After this operation, the model will have learned patterns from the training data.
pipeline.fit(X_train, y_train)

# Make predictions on the test data using the trained machine learning pipeline.
# - This step uses the trained Linear Regression model to predict target values (y_pred)
#   based on the input features of the test data (X_test).
y_pred = pipeline.predict(X_test)

y_pred

mse = mean_squared_error(y_test, y_pred)
print(f"Mean Squared Error: {mse}")

# Calculate R-squared (coefficient of determination)
r_squared = r2_score(y_test, y_pred)
print(f"R-squared: {r_squared}")

# Calculate Mean Absolute Error
mae = mean_absolute_error(y_test, y_pred)
print(f"Mean Absolute Error: {mae}")

## Decision Tree Regression

#  Initialize the DecisionTreeRegressor with a specific random state
tree_regressor = DecisionTreeRegressor(random_state=42)

# Fit the DecisionTreeRegressor model to the training data
tree_regressor.fit(X_train, y_train)


# Predict the target variable on the test data
y_pred = tree_regressor.predict(X_test)

y_pred

# Calculate the Mean Squared Error (MSE)
mse = mean_squared_error(y_test, y_pred)
print("Mean Squared Error:", mse)

# Calculate the Mean Absolute Error (MAE)
mae = mean_absolute_error(y_test, y_pred)
print("Mean Absolute Error:", mae)

# Calculate the R-squared score
r2 = r2_score(y_test, y_pred)
print("R-squared:", r2)

## Random Forest Regressor

# Create a Random Forest Regressor model
rf_regressor = RandomForestRegressor(n_estimators=100, random_state=42)

# Fit the model on the training data
rf_regressor.fit(X_train, y_train)

# Make predictions on the test data
y_pred_rf = rf_regressor.predict(X_test)

# Save the trained model to a pkl file
joblib.dump(rf_regressor, '/content/drive/MyDrive/CSCT MASTER PROJECT/random_forest_model.pkl')

y_pred_rf

# Calculate Mean Absolute Error (MAE)
mae_rf = mean_absolute_error(y_test, y_pred_rf)
print("Mean Absolute Error:", mae_rf)

# Calculate R-squared score
r2_rf = r2_score(y_test, y_pred_rf)
print("R-squared:", r2_rf)

# Calculate Mean Squared Error (MSE)
mse_rf = mean_squared_error(y_test, y_pred_rf)
print("Mean Squared Error:", mse_rf)
