==> Step_4_User_Interface[1].ipynb <==
import tkinter as tk
from tkinter import ttk
import joblib
import pandas as pd
from sklearn.preprocessing import LabelEncoder  # Add this import

!pip install pyvirtualdisplay

!apt-get install -y xvfb

from pyvirtualdisplay import Display
from tkinter import Tk, Label

# Start a virtual display
display = Display(visible=0, size=(800, 600))
display.start()

import tkinter as tk
from tkinter import ttk
import joblib
import pandas as pd
from sklearn.preprocessing import LabelEncoder  # Add this import
from sklearn.ensemble import RandomForestRegressor

# Load your trained Random Forest model from the pkl file
rf_regressor = joblib.load('/content/drive/MyDrive/CSCT MASTER PROJECT/random_forest_model.pkl')

# Create a function to predict yield
def predict_yield():
    state = state_var.get()
    district = district_var.get()
    crop = crop_var.get()
    year = int(year_var.get())
    season = season_var.get().strip()
    area = float(area_var.get())


    # Create and fit new LabelEncoder instances for encoding
    label_encoder_state = LabelEncoder()
    label_encoder_district = LabelEncoder()
    label_encoder_crop = LabelEncoder()
    label_encoder_season = LabelEncoder()
    label_encoder_area_units = LabelEncoder()

    # Apply LabelEncoder to each categorical input
    state_encoded = label_encoder_state.fit_transform([state])
    district_encoded = label_encoder_district.fit_transform([district])
    crop_encoded = label_encoder_crop.fit_transform([crop])
    season_encoded = label_encoder_season.fit_transform([season])
    area_units_encoded = label_encoder_area_units.fit_transform([area_unit_var.get().strip()])

    features = pd.DataFrame({
        'State': state_encoded,
        'District': district_encoded,
        'Crop': crop_encoded,
        'Year': [year],
        'Season': [season_encoded],
        'Area': [area],
        'Area Units': area_units_encoded
    })

      # Perform prediction using the model
    prediction = rf_regressor.predict(features)

    # Display the prediction in a label
    result_label.config(text=f"Predicted Value: {prediction[0]}")


            # Load state-district mapping
state_district_mapping = {
'Andaman and Nicobar Islands': ['NICOBARS' , 'NORTH AND MIDDLE ANDAMAN' , 'SOUTH ANDAMANS' , 'Andaman and Nicobar Islands'],
'Andhra Pradesh': ['ANANTAPUR' , 'EAST GODAVARI' , 'KRISHNA' , 'VIZIANAGARAM' , 'WEST GODAVARI' , 'ADILABAD' , 'CHITTOOR',
                   'GUNTUR' , 'KADAPA' , 'KARIMNAGAR' , 'KHAMMAM' , 'KURNOOL' , 'MAHBUBNAGAR' , 'MEDAK' , 'NALGONDA',
                   'NIZAMABAD' , 'PRAKASAM' , 'RANGAREDDI','SPSR NELLORE' , 'SRIKAKULAM' , 'VISAKHAPATANAM' , 'WARANGAL',
                   'HYDERABAD'],
'Arunachal Pradesh': ['CHANGLANG' , 'DIBANG VALLEY' , 'EAST KAMENG' , 'EAST SIANG' , 'KURUNG KUMEY' , 'LOHIT' ,
                      'LOWER DIBANG VALLEY' , 'LOWER SUBANSIRI' , 'PAPUM PARE' , 'TAWANG' , 'TIRAP' , 'UPPER SIANG' ,
                       'UPPER SUBANSIRI' , 'WEST KAMENG' , 'WEST SIANG',
 'ANJAW' , 'KRA DAADI' , 'LONGDING' , 'NAMSAI' , 'SIANG' , 'KAMLE' , 'LOWER SIANG' , 'LEPARADA' , 'PAKKE KESSANG' , 'SHI YOMI'],
'Assam': ['BARPETA' , 'BONGAIGAON' , 'CACHAR' , 'DARRANG' , 'DHEMAJI' , 'DHUBRI' , 'DIBRUGARH' , 'DIMA HASAO' , 'GOALPARA' ,
          'GOLAGHAT' , 'HAILAKANDI' , 'JORHAT' , 'KAMRUP' , 'KARBI ANGLONG' , 'KARIMGANJ' , 'KOKRAJHAR' , 'LAKHIMPUR' ,
          'MARIGAON' , 'NAGAON','NALBARI' , 'SIVASAGAR' , 'SONITPUR' , 'TINSUKIA' , 'BAKSA' , 'CHIRANG' , 'KAMRUP METRO' ,
          'UDALGURI' , 'BISWANATH' , 'CHARAIDEO' , 'HOJAI' , 'MAJULI' , 'SOUTH SALMARA MANCACHAR' , 'WEST KARBI ANGLONG'],
'Bihar': ['ARARIA' , 'ARWAL' , 'AURANGABAD' , 'BANKA' , 'BEGUSARAI' , 'BHAGALPUR' , 'BHOJPUR' , 'BUXAR' , 'DARBHANGA' , 'GAYA' ,
          'GOPALGANJ' , 'JAMUI' , 'JEHANABAD' , 'KAIMUR (BHABUA)' , 'KATIHAR' , 'KHAGARIA' , 'KISHANGANJ' , 'LAKHISARAI' ,
          'MADHEPURA' , 'MADHUBANI' , 'MUNGER' , 'MUZAFFARPUR' , 'NALANDA' , 'NAWADA' , 'PASHCHIM CHAMPARAN' , 'PATNA' ,
          'PURBI CHAMPARAN' , 'PURNIA' , 'ROHTAS' , 'SAHARSA' , 'SAMASTIPUR' , 'SARAN' , 'SHEIKHPURA' , 'SHEOHAR' ,
          'SITAMARHI' , 'SIWAN' , 'SUPAUL' , 'VAISHALI' , 'BOKARO' , 'CHATRA' , 'DEOGHAR' , 'DHANBAD' , 'DUMKA' ,
          'EAST SINGHBUM' , 'GARHWA' , 'GIRIDIH' , 'GODDA' , 'GUMLA' , 'HAZARIBAGH' , 'KODERMA' , 'LOHARDAGA' ,
          'PAKUR' , 'PALAMU' , 'RANCHI' , 'SAHEBGANJ' , 'WEST SINGHBHUM'],
'Chandigarh': ['CHANDIGARH'],
'Chhattisgarh': ['BASTAR' , 'BILASPUR' , 'DANTEWADA' , 'DHAMTARI' , 'DURG' , 'JANJGIR-CHAMPA''JASHPUR' , 'KABIRDHAM' ,
                 'KANKER' , 'KORBA' , 'KOREA' , 'MAHASAMUND' , 'RAIGARH' , 'RAIPUR' , 'RAJNANDGAON' , 'SURGUJA' , 'NARAYANPUR' ,
                 'BIJAPUR' , 'BALOD' , 'BALODA BAZAR' , 'BALRAMPUR' , 'BEMETARA' , 'GARIYABAND' , 'KONDAGAON' , 'MUNGELI' ,
                 'SUKMA' , 'SURAJPUR' , 'GAURELLA-PENDRA-MARWAHI'],
'Dadra and Nagar Haveli': ['DADRA AND NAGAR HAVELI'],
'Daman and Diu': ['Daman and Diu' , 'DIU' , 'DAMAN'],
'Delhi': ['Delhi' , 'DELHI_TOTAL'],
'Goa': ['Goa' , 'NORTH GOA' , 'SOUTH GOA'],
'Gujarat': ['AHMADABAD' , 'AMRELI' , 'ANAND' , 'BANAS KANTHA' , 'BHARUCH' , 'BHAVNAGAR' , 'DANG' , 'DOHAD' , 'GANDHINAGAR' ,
            'JAMNAGAR' , 'JUNAGADH' , 'KHEDA' , 'MAHESANA' , 'NARMADA' , 'NAVSARI' , 'PANCH MAHALS' , 'PATAN' , 'PORBANDAR' ,
            'RAJKOT' , 'SABAR KANTHA','SURAT' , 'SURENDRANAGAR' , 'VADODARA' , 'VALSAD' , 'KACHCHH' , 'TAPI' , 'ARAVALLI' ,
            'BOTAD' , 'CHHOTAUDEPUR' , 'GIR SOMNATH' , 'MAHISAGAR' , 'DEVBHUMI DWARKA' , 'MORBI'],
'Haryana': ['AMBALA' , 'BHIWANI' , 'FARIDABAD' , 'FATEHABAD' , 'GURGAON' , 'HISAR' , 'JHAJJAR' , 'JIND' , 'KAITHAL' ,
            'KARNAL' , 'KURUKSHETRA' , 'MAHENDRAGARH' , 'PANCHKULA' , 'PANIPAT' , 'REWARI' , 'ROHTAK' , 'SIRSA' ,
            'SONIPAT' , 'YAMUNANAGAR' , 'MEWAT','PALWAL' , 'CHARKI DADRI'],
'Himachal Pradesh': ['BILASPUR' , 'KANGRA' , 'KULLU' , 'MANDI' , 'SHIMLA' , 'SOLAN' , 'UNA' , 'CHAMBA' , 'HAMIRPUR' ,
                     'SIRMAUR' , 'KINNAUR' , 'LAHUL AND SPITI'],
'Jammu and Kashmir': ['DODA' , 'JAMMU' , 'KATHUA' , 'RAJAURI' , 'UDHAMPUR' , 'KARGIL' , 'LEH LADAKH' , 'SRINAGAR' ,
                      'BADGAM' , 'BARAMULLA' , 'POONCH' , 'PULWAMA' , 'ANANTNAG' , 'KUPWARA' , 'REASI' , 'SAMBA' ,
                      'KISHTWAR' , 'RAMBAN' , 'KULGAM' , 'BANDIPORA' , 'GANDERBAL','SHOPIAN'],
'Jharkhand': ['CHATRA' , 'DUMKA' , 'GARHWA' , 'GODDA' , 'GUMLA' , 'HAZARIBAGH' , 'KODERMA' , 'LATEHAR' , 'LOHARDAGA' , 'PAKUR' ,
              'PALAMU' , 'RANCHI' , 'SAHEBGANJ' , 'SARAIKELA KHARSAWAN' , 'SIMDEGA' , 'WEST SINGHBHUM' , 'BOKARO' , 'DEOGHAR' ,
              'DHANBAD' , 'EAST SINGHBUM' , 'GIRIDIH' , 'JAMTARA' , 'KHUNTI' , 'RAMGARH'],
'Karnataka': ['BAGALKOT' , 'BANGALORE RURAL' , 'BELGAUM' , 'BELLARY' , 'BENGALURU URBAN' , 'CHAMARAJANAGAR' , 'CHIKMAGALUR' ,
              'CHITRADURGA' , 'DAKSHIN KANNAD' , 'DAVANGERE' , 'DHARWAD' , 'GADAG' , 'HASSAN' , 'HAVERI' , 'KODAGU' , 'KOLAR' ,
              'MANDYA' , 'MYSORE','SHIMOGA' , 'TUMKUR' , 'UDUPI' , 'UTTAR KANNAD' , 'BIDAR' , 'BIJAPUR' , 'GULBARGA' , 'KOPPAL',
              'RAICHUR' , 'RAMANAGARA' , 'CHIKBALLAPUR' , 'YADGIR' , 'BAGALKOTE' , 'BALLARI' , 'BELAGAVI' , 'CHAMARAJANAGARA' ,
              'CHIKKABALLAPURA' , 'CHIKKAMAGALURU' , 'DAKSHINA KANNADA' , 'MYSURU' , 'SHIVAMOGGA' , 'TUMAKURU' ,
              'UTTARA KANNADA' , 'VIJAYAPURA' , 'KALABURAGI' , 'YADAGIRI'],
'Kerala': ['ALAPPUZHA' , 'ERNAKULAM' , 'IDUKKI' , 'KANNUR' , 'KASARAGOD' , 'KOLLAM' , 'KOTTAYAM' , 'KOZHIKODE' , 'MALAPPURAM' ,
           'PALAKKAD' , 'PATHANAMTHITTA' , 'THIRUVANANTHAPURAM' , 'THRISSUR' , 'WAYANAD'],
'Laddakh': ['KARGIL'],
'Madhya Pradesh': ['ANUPPUR' , 'ASHOKNAGAR' , 'BALAGHAT' , 'BARWANI' , 'BETUL' , 'BHIND' , 'BHOPAL' , 'BURHANPUR' ,
                   'CHHATARPUR' , 'CHHINDWARA' , 'DAMOH' , 'DATIA' , 'DEWAS' , 'DHAR' , 'DINDORI' , 'GUNA' , 'GWALIOR' ,
                   'HARDA' , 'HOSHANGABAD' , 'INDORE' , 'JABALPUR','JHABUA' , 'KATNI' , 'KHANDWA' , 'KHARGONE' , 'MANDLA' ,
                   'MANDSAUR' , 'MORENA','NARSINGHPUR' , 'NEEMUCH' , 'PANNA' , 'RAISEN' , 'RAJGARH' , 'RATLAM' , 'REWA' ,
                   'SAGAR' , 'SATNA' , 'SEHORE' , 'SEONI' , 'SHAHDOL' , 'SHAJAPUR' , 'SHEOPUR' , 'SHIVPURI' , 'SIDHI' ,
                   'TIKAMGARH' , 'UJJAIN' , 'UMARIA' , 'VIDISHA' , 'ALIRAJPUR' , 'SINGRAULI' , 'AGAR MALWA' , 'NIWARI' ,
                   'BASTAR' , 'BILASPUR' , 'DANTEWADA' , 'DHAMTARI' , 'DURG' , 'JANJGIR-CHAMPA' , 'JASHPUR' , 'KABIRDHAM' ,
                   'KANKER' , 'KORBA' , 'KOREA' , 'MAHASAMUND' , 'RAIPUR' , 'RAJNANDGAON' , 'SURGUJA'],
'Maharashtra': ['AHMEDNAGAR' , 'AKOLA' , 'AMRAVATI' , 'AURANGABAD' , 'BEED' , 'BHANDARA' , 'BULDHANA' , 'CHANDRAPUR' ,
                'DHULE' , 'GADCHIROLI' , 'GONDIA' , 'HINGOLI' , 'JALGAON' , 'JALNA','KOLHAPUR' , 'LATUR' , 'NAGPUR' , 'NANDED' ,
                'NANDURBAR' , 'NASHIK' , 'OSMANABAD' , 'PARBHANI' , 'PUNE' , 'RAIGAD' , 'RATNAGIRI' , 'SANGLI' , 'SATARA' ,
                'SOLAPUR' , 'THANE' , 'WARDHA' , 'WASHIM' , 'YAVATMAL' , 'SINDHUDURG' , 'PALGHAR' , 'MUMBAI SUBURBAN' ,
                'MUMBAI'],
'Manipur': ['SENAPATI' , 'BISHNUPUR' , 'CHANDEL' , 'CHURACHANDPUR' , 'IMPHAL EAST' , 'IMPHAL WEST' , 'TAMENGLONG' , 'THOUBAL' ,
            'UKHRUL'],
'Meghalaya': ['EAST GARO HILLS' , 'EAST JAINTIA HILLS' , 'EAST KHASI HILLS' , 'RI BHOI' , 'SOUTH GARO HILLS' , 'WEST GARO HILLS',
              'WEST KHASI HILLS' , 'NORTH GARO HILLS' , 'SOUTH WEST GARO HILLS' , 'SOUTH WEST KHASI HILLS' ,
              'WEST JAINTIA HILLS'],
'Mizoram': ['AIZAWL' , 'CHAMPHAI' , 'KOLASIB' , 'LUNGLEI' , 'MAMIT' , 'SAIHA' , 'LAWNGTLAI' , 'SERCHHIP'],
'Nagaland': ['DIMAPUR' , 'KOHIMA' , 'MOKOKCHUNG' , 'MON' , 'PHEK' , 'TUENSANG' , 'WOKHA' , 'ZUNHEBOTO' , 'KIPHIRE' ,
             'LONGLENG' , 'PEREN'],
'Odisha': ['ANUGUL' , 'BALANGIR' , 'BALESHWAR' , 'BARGARH' , 'BHADRAK' , 'BOUDH' , 'CUTTACK' , 'DEOGARH' ,
           'DHENKANAL' , 'GAJAPATI' , 'GANJAM' , 'JAGATSINGHAPUR' , 'JAJAPUR' , 'JHARSUGUDA' , 'KALAHANDI' ,
           'KANDHAMAL' , 'KENDRAPARA' , 'KENDUJHAR' , 'KHORDHA' , 'KORAPUT' , 'MALKANGIRI' , 'MAYURBHANJ' , 'NABARANGPUR' ,
           'NAYAGARH' , 'NUAPADA' , 'RAYAGADA' , 'SAMBALPUR' , 'SONEPUR' , 'SUNDARGARH' , 'PURI'],
'Puducherry': ['MAHE' , 'PONDICHERRY' , 'KARAIKAL' , 'YANAM'],
'Punjab': ['AMRITSAR' , 'BATHINDA' , 'FARIDKOT' , 'FATEHGARH SAHIB' , 'FIROZEPUR' , 'HOSHIARPUR' , 'JALANDHAR' , 'KAPURTHALA' ,
           'LUDHIANA' , 'MOGA' , 'MUKTSAR','NAWANSHAHR' , 'PATIALA' , 'RUPNAGAR' , 'SANGRUR' , 'MANSA' , 'GURDASPUR' , 'S' ,
           'TARN TARAN' , 'BARNALA' , 'FAZILKA' , 'PATHANKOT' , 'SHAHID BHAGAT SINGH NAGAR'],
'Rajasthan': ['AJMER' , 'ALWAR' , 'BANSWARA' , 'BARAN' , 'BHARATPUR' , 'BHILWARA' , 'BIKANER' , 'BUNDI' , 'CHITTORGARH' ,
              'DAUSA' , 'DHOLPUR' , 'DUNGARPUR' , 'GANGANAGAR' , 'HANUMANGARH' , 'JAIPUR' , 'JAISALMER' , 'JHALAWAR' ,
              'KARAULI' , 'KOTA' , 'NAGAUR' , 'PALI' , 'RAJSAMAND' , 'SAWAI MADHOPUR' , 'SIKAR' , 'SIROHI' , 'TONK' ,
              'UDAIPUR' , 'BARMER' , 'CHURU' , 'JALORE' , 'JHUNJHUNU' , 'JODHPUR' , 'PRATAPGARH'],
'Sikkim': ['EAST DISTRICT' , 'NORTH DISTRICT' , 'SOUTH DISTRICT' , 'WEST DISTRICT'],
'Tamil Nadu': ['COIMBATORE' , 'CUDDALORE' , 'DHARMAPURI' , 'DINDIGUL' , 'ERODE' , 'KANCHIPURAM' , 'KARUR' ,
               'MADURAI' , 'NAMAKKAL' , 'PERAMBALUR' , 'PUDUKKOTTAI' , 'RAMANATHAPURAM' , 'SALEM' , 'SIVAGANGA' ,
               'THANJAVUR' , 'THENI' , 'THIRUVALLUR' , 'THIRUVARUR' , 'THOOTHUKUDI' , 'TIRUCHIRAPPALLI' , 'TIRUNELVELI' ,
               'TIRUVANNAMALAI' , 'VELLORE' , 'VILLUPURAM' , 'VIRUDHUNAGAR' , 'NAGAPATTINAM' , 'THE NILGIRIS' ,
               'KANNIYAKUMARI' , 'KRISHNAGIRI' , 'TIRUPPUR' , 'ARIYALUR' , 'TUTICORIN' , 'CHENNAI' , 'KALLAKURICHI',
 'TENKASI' , 'TIRUPATHUR' , 'CHENGALPATTU' , 'RANIPET'],
'Telangana': ['ADILABAD' , 'KARIMNAGAR' , 'KHAMMAM' , 'MAHBUBNAGAR' , 'MEDAK' , 'NALGONDA' , 'NIZAMABAD' , 'RANGAREDDI' ,
              'WARANGAL' , 'BHADRADRI' , 'JAGITIAL' , 'JANGOAN' , 'JAYASHANKAR' , 'JOGULAMBA' , 'KAMAREDDY' ,
              'KOMARAM BHEEM ASIFABAD' , 'MAHABUBABAD' , 'MANCHERIAL' , 'MEDCHAL' , 'MULUGU' , 'NAGARKURNOOL' ,
              'NARAYANAPET' , 'NIRMAL' , 'PEDDAPALLI' , 'RAJANNA' , 'SANGAREDDY' , 'SIDDIPET' , 'SURYAPET' , 'VIKARABAD' ,
              'WANAPARTHY' , 'WARANGAL URBAN' , 'YADADRI' , 'HANUMAKONDA' , 'MEDCHAL MALKAJGIRI' , 'DADRA AND NAGAR HAVELI' ,
              'DIU' , 'DAMAN'],
'Tripura': ['DHALAI' , 'NORTH TRIPURA' , 'SOUTH TRIPURA' , 'WEST TRIPURA' , 'GOMATI' , 'KHOWAI' , 'SEPAHIJALA' , 'UNAKOTI'],
'Uttar Pradesh': ['AGRA' , 'ALIGARH' , 'ALLAHABAD' , 'AMBEDKAR NAGAR' , 'AMROHA' , 'AURAIYA' , 'AZAMGARH' , 'BAGHPAT' ,
                  'BAHRAICH' , 'BALLIA' , 'BALRAMPUR' , 'BANDA' , 'BARABANKI','BAREILLY' , 'BASTI' , 'BIJNOR' , 'BUDAUN' ,
                  'BULANDSHAHR' , 'CHANDAULI' , 'CHITRAKOOT' , 'DEORIA' , 'ETAH' , 'ETAWAH' , 'FAIZABAD' , 'FARRUKHABAD' ,
                  'FATEHPUR' , 'FIROZABAD' , 'GAUTAM BUDDHA NAGAR' , 'GHAZIABAD' , 'GHAZIPUR' , 'GONDA' , 'GORAKHPUR' ,
                  'HAMIRPUR' , 'HARDOI' , 'HATHRAS' , 'JALAUN' , 'JAUNPUR' , 'JHANSI' , 'KANNAUJ' , 'KANPUR DEHAT' ,
                  'KANPUR NAGAR' , 'KAUSHAMBI' , 'KHERI' , 'KUSHI NAGAR' , 'LALITPUR' , 'LUCKNOW' , 'MAHARAJGANJ' ,
                  'MAHOBA' , 'MAINPURI' , 'MATHURA' , 'MAU' , 'MEERUT' , 'MIRZAPUR' , 'MORADABAD' , 'MUZAFFARNAGAR' ,
                  'PILIBHIT' , 'PRATAPGARH' , 'RAE BARELI' , 'RAMPUR' , 'SAHARANPUR' , 'SANT KABEER NAGAR' ,
                  'SANT RAVIDAS NAGAR' , 'SHAHJAHANPUR' , 'SHRAVASTI' , 'SIDDHARTH NAGAR' , 'SITAPUR' , 'SONBHADRA' ,
                  'SULTANPUR' , 'UNNAO' , 'VARANASI' , 'KASGANJ' , 'AMETHI','HAPUR' , 'SAMBHAL' , 'SHAMLI' , 'ALMORA' ,
                  'CHAMOLI' , 'DEHRADUN' , 'HARIDWAR' , 'NAINITAL' , 'PAURI GARHWAL' , 'PITHORAGARH' , 'TEHRI GARHWAL' ,
                  'UDAM SINGH NAGAR' , 'UTTAR KASHI' , 'BAGESHWAR' , 'CHAMPAWAT' , 'RUDRA PRAYAG'],
'Uttarakhand': ['CHAMOLI' , 'CHAMPAWAT' , 'DEHRADUN' , 'HARIDWAR' , 'NAINITAL' , 'PAURI GARHWAL' , 'PITHORAGARH' ,
                'RUDRA PRAYAG' , 'TEHRI GARHWAL' , 'UDAM SINGH NAGAR' , 'UTTAR KASHI' , 'ALMORA' , 'BAGESHWAR'],
'West Bengal': ['24 PARAGANAS NORTH' , '24 PARAGANAS SOUTH' , 'BANKURA' , 'BIRBHUM' , 'COOCHBEHAR' , 'DARJEELING' ,
                'DINAJPUR DAKSHIN' , 'DINAJPUR UTTAR' , 'HOOGHLY' , 'HOWRAH' , 'JALPAIGURI' , 'MALDAH' , 'MEDINIPUR EAST' ,
                'MEDINIPUR WEST' , 'MURSHIDABAD' , 'NADIA' , 'PURBA BARDHAMAN' , 'PURULIA' , 'ALIPURDUAR' , 'JHARGRAM' ,
                'PASCHIM BARDHAMAN' , 'KALIMPONG']
}

# Create the main tkinter window
root = tk.Tk()
root.title("Crop Yield Prediction")
root.configure(bg="White")

# Create a bold and larger font for the title
title_font = ("Helvetica", 20, "bold")
label_font = ("Helvetica", 10, "bold")

# Create a label for the title
title_label = ttk.Label(root, text="Welcome to Yield Predictor", font=title_font)
title_label.pack()

# Create a frame to hold the input fields
input_frame = ttk.Frame(root, padding=20)
input_frame.pack()

# Get the first state from the keys of the state_district_mapping dictionary
default_state = next(iter(state_district_mapping.keys()))

# Create dropdown menu for State
state_frame = ttk.Frame(root)
state_frame.pack(fill="both", padx=10, pady=10)
state_label = ttk.Label(state_frame, text="Select State:", background="White", font=label_font)
state_label.pack(side="left")

# Set the default state as the default value
state_var = tk.StringVar(value=default_state)
state_menu = ttk.OptionMenu(state_frame, state_var, *state_district_mapping.keys())
state_menu.pack(side="left")

def update_district_options(*args):
    selected_state = state_var.get()
    if selected_state in state_district_mapping:
        district_menu["menu"].delete(0, "end")
        for district in state_district_mapping[selected_state]:
            district_menu["menu"].add_command(label=district, command=tk._setit(district_var, district))

# Create a trace on state_var to call update_district_options() when the state is changed
state_var.trace('w', update_district_options)


# Create dropdown menu for District
district_frame = ttk.Frame(root)
district_frame.pack(fill="both", padx=10, pady=10)
district_label = ttk.Label(district_frame, text="Select District:", background="White", font=label_font)
district_label.pack(side="left")
district_var = tk.StringVar()
district_menu = ttk.OptionMenu(district_frame, district_var, "")
district_menu.pack(side="left")



# Create dropdown menu for Crop
crop_frame = ttk.Frame(root)
crop_frame.pack(fill="both", padx=10, pady=10)
crop_label = ttk.Label(crop_frame, text="Select Crop:", background="White", font=label_font)
crop_label.pack(side="left")
crop_options = ['Arecanut' , 'Banana' , 'Black pepper' , 'Cashewnut' , 'Coconut' , 'Dry chillies' , 'Ginger' , 'Other Kharif pulses' , 'other oilseeds' , 'Rice' , 'Sugarcane' , 'Sweet potato' , 'Arhar/Tur' , 'Bajra' , 'Castor seed' , 'Coriander' , 'Cotton(lint)' , 'Gram' , 'Groundnut' , 'Horse-gram' , 'Jowar' , 'Linseed' , 'Maize' , 'Mesta' , 'Moong(Green Gram)' , 'Niger seed' , 'Onion' , 'Other Rabi pulses' , 'Potato' , 'Ragi' , 'Rapeseed &Mustard' , 'Safflower' , 'Sesamum' , 'Small millets' , 'Soyabean' , 'Sunflower' , 'Tapioca' , 'Tobacco' , 'Turmeric' , 'Urad' , 'Wheat' , 'Oilseeds total' , 'Jute' , 'Masoor' , 'Peas & beans (Pulses)' , 'Barley'
 'Garlic' , 'Khesari' , 'Sannhamp' , 'Guar seed' , 'Moth' , 'Cardamom' , 'Other Cereals' , 'Cowpea(Lobia)' ,
                'Dry Ginger' , 'Other Summer Pulses']  # Replace with your actual crop options
crop_var = tk.StringVar(value=crop_options[0])
crop_menu = ttk.OptionMenu(crop_frame, crop_var, *crop_options)
crop_menu.pack(side="left")

# Create dropdown menu for Year
year_frame = ttk.Frame(root)
year_frame.pack(fill="both", padx=10, pady=10)

year_label = ttk.Label(year_frame, text="Select Year:", background="White", font=label_font)
year_label.pack(side="left")
year_options =  [2023, 2024, 2025,2026,2027]
year_var = tk.StringVar(value=year_options[0])
year_menu = ttk.OptionMenu(year_frame, year_var, *year_options)
year_menu.pack(side="left")

# Create dropdown menu for Season
season_frame = ttk.Frame(root)
season_frame.pack(fill="both", padx=10, pady=10)
season_label = ttk.Label(season_frame, text="Select Season:", background="White", font=label_font)
season_label.pack(side="left")
season_options =  ['Kharif' , 'Whole Year' , 'Rabi' , 'Autumn' , 'Summer' , 'Winter'] # Replace with your actual season options
season_var = tk.StringVar(value=season_options[0])
season_menu = ttk.OptionMenu(season_frame, season_var, *season_options)
season_menu.pack(side="left")

# Create entry for Area
area_frame = ttk.Frame(root)
area_frame.pack(fill="both", padx=10, pady=10)
area_label = ttk.Label(area_frame, text="Enter Area:", background="White", font=label_font)
area_label.pack(side="left")
area_var = tk.StringVar()
area_entry = ttk.Entry(area_frame, textvariable=area_var)
area_entry.pack(side="left")


# Create dropdown menu for Area Unit
area_unit_frame = ttk.Frame(root)
area_unit_frame.pack(fill="both", padx=10, pady=10)
area_unit_label = ttk.Label(area_unit_frame, text="Select Area Unit:", background="White", font=label_font)
area_unit_label.pack(side="left")
area_units_options = ['Hectares']
area_unit_var = tk.StringVar(value=area_units_options[0])
area_unit_menu = ttk.OptionMenu(area_unit_frame, area_unit_var, *area_units_options)
area_unit_menu.pack(side="left")

# Create a prediction button with blue background and white text color
prediction_button = ttk.Button(root, text="Calculate Predicted Yield",command=predict_yield)
prediction_button.pack()

# Configure the style for the prediction button
style = ttk.Style()
style.configure("PredictionButton.TButton", background="blue")

prediction_button.configure(style="PredictionButton.TButton")

# Display the prediction result
result_label = ttk.Label(root, text="")
result_label.pack()

# Start the tkinter event loop
root.mainloop()

# Stop the virtual display when excecution done
display.stop()
